sepolicy: Added rules as per new sensorservice
In new AOSP sensor manager needs to be initialized with a name.
Sensor service is getting launched directly rather than through a
thread as earlier. Sensorservice is not being allowed to find the
list of services from service manager and that caused camera crash.

Added rule for sensorservice to find the list of services.

Change-Id: I00521b91deba8dc364e6601ab45c2a217a5f9cf6