SEandroid: fingerprint: add QFP daemon and sensor lib permissions
1. Sensors wake-up library runs in the system_server context.
The wake-up library and the QFP daemon communicate through binder.
This change gives the system_server and the QFP daemon permission
to access this binder.
2. Adds permission to system_server (wake-up library) to execute shell
commands. The wake-up library needs to call a shell command to vibrate
the device after finger detection and match failure.
3. Allow QFP daemon permission to sensors socket to enable QFP daemon to
communicate with the SSC through the sensors process.

Change-Id: Ieb79a6ade856490772623d7d331ea17ad2f0b149

Conflicts:
	common/system_server.te